Tests/LibWeb/Text/expected/HTML/Window-set-properties.txt

window.__NEXT_DATA__ = [object HTMLScriptElement]
window.__NEXT_DATA__ = undefined
window[0] = undefined
e = TypeError: Cannot set property '0' of [object WindowProxy]
window[0] = undefined
window[0] = 42
